Revert "Revert "Rename App Data terminology to Fields""
