+++ lib/models/message_template.dart
@@ -334,10 +334,10 @@ class MessageTemplate extends HiveObject {
 
       if (field is Map<String, dynamic>) {
         fieldName = field['fieldName'] as String? ?? '';
-        category = field['category'] as String? ?? 'Custom Fields';
+        category = field['category'] as String? ?? 'Fields';
       } else {
         fieldName = field.fieldName as String? ?? '';
-        category = field.category as String? ?? 'Custom Fields';
+        category = field.category as String? ?? 'Fields';
       }
 
       if (fieldName.isNotEmpty) {
