+++ lib/models/pdf_template.dart
@@ -229,7 +229,7 @@ class PDFTemplate extends HiveObject {
         'notes', 'terms', 'upgradeQuoteText'
       ],
 
-      'Custom Fields': [
+      'Fields': [
         'customText1', 'customText2', 'customText3',
         'customNumeric1', 'customNumeric2',
         'customDate1', 'customDate2',
@@ -312,7 +312,7 @@ class PDFTemplate extends HiveObject {
         'contact': 'Contact Information', // Will create new if doesn't exist
         'legal': 'Legal Information',
         'pricing': 'Pricing Information',
-        'custom': 'Custom App Data',
+        'custom': 'Fields',
       };
 
       // Process each custom field category
