+++ lib/screens/template_editor_screen.dart
@@ -479,6 +479,13 @@ class _TemplateEditorScreenState extends State<TemplateEditorScreen> {
                 icon: const Icon(Icons.close),
                 onPressed: () => Navigator.pop(context),
               ),
+              actions: [
+                IconButton(
+                  icon: const Icon(Icons.help_outline),
+                  tooltip: 'Mapping help',
+                  onPressed: _showMappingHelp,
+                ),
+              ],
             ),
             body: _buildAppDataFieldsList(pdfFieldInfo),
           ),
@@ -680,7 +687,7 @@ class _TemplateEditorScreenState extends State<TemplateEditorScreen> {
       pdfFormFieldName: pdfFieldName,
       detectedPdfFieldType: PdfFormFieldType.values.firstWhere(
             (e) => e.toString() == pdfFieldInfo['type'],
-        orElse: () => PdfFormFieldType. unknown,
+        orElse: () => PdfFormFieldType.unknown,
       ),
       pageNumber: pdfFieldInfo['page'] as int,
     );
@@ -715,7 +722,7 @@ class _TemplateEditorScreenState extends State<TemplateEditorScreen> {
 
     setState(() {
       mapping.pdfFormFieldName = '';
-      mapping.detectedPdfFieldType = PdfFormFieldType. unknown;
+      mapping.detectedPdfFieldType = PdfFormFieldType.unknown;
       mapping.visualX = null;
       mapping.visualY = null;
       mapping.visualWidth = null;
@@ -727,6 +734,25 @@ class _TemplateEditorScreenState extends State<TemplateEditorScreen> {
       const SnackBar(
         content: Text("Field mapping removed."),
         backgroundColor: Colors.orange,
+      ),
+    );
+  }
+
+  void _showMappingHelp() {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('How to Map Fields'),
+        content: const Text(
+          'Select a field from the list to link it with the chosen PDF field. '
+          'Existing links can be replaced by choosing another field.',
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
       ),
     );
   }
